Create mixin for route view handling and use in SearchInterface and  galeries page; add SearchViewToggles to galleries page
