+++ packages/portal/src/pages/galleries/_.vue
@@ -156,6 +156,10 @@
             >
               {{ displayItemCount }}
             </h2>
+            <SearchViewToggles
+              v-model="view"
+              class="ml-auto"
+            />
           </b-col>
         </b-row>
         <b-row>
@@ -167,6 +171,7 @@
                     :items="set.items"
                     :show-pins="setIsEntityBestItems && userIsEntityEditor"
                     :user-editable-items="userCanEditSet"
+                    :view="view"
                     @endItemDrag="reorderItems"
                   />
                 </b-col>
@@ -194,9 +199,11 @@
   } from '@/plugins/europeana/data';
   import { langMapValueForLocale } from '@europeana/i18n';
   import ItemPreviewCardGroup from '@/components/item/ItemPreviewCardGroup';
+  import SearchViewToggles from '@/components/search/SearchViewToggles.vue';
   import ShareButton from '@/components/share/ShareButton.vue';
   import ShareSocialModal from '@/components/share/ShareSocialModal.vue';
   import entityBestItemsSetMixin from '@/mixins/europeana/entities/entityBestItemsSet';
+  import itemPreviewCardGroupViewMixin from '@/mixins/europeana/item/itemPreviewCardGroupView';
   import langAttributeMixin from '@/mixins/langAttribute';
   import pageMetaMixin from '@/mixins/pageMeta';
   import redirectToMixin from '@/mixins/redirectTo';
@@ -208,6 +215,7 @@
       LoadingSpinner: () => import('@/components/generic/LoadingSpinner'),
       ErrorMessage: () => import('@/components/error/ErrorMessage'),
       ItemPreviewCardGroup,
+      SearchViewToggles,
       ShareButton,
       ShareSocialModal,
       SetFormModal: () => import('@/components/set/SetFormModal'),
@@ -218,6 +226,7 @@
     },
     mixins: [
       entityBestItemsSetMixin,
+      itemPreviewCardGroupViewMixin,
       langAttributeMixin,
       redirectToMixin,
       pageMetaMixin
