+++ packages/portal/tests/unit/components/search/SearchInterface.spec.js
@@ -54,7 +54,6 @@ const factory = ({ $fetchState = {}, mocks = {}, propsData = {}, data = {} } = {
     $store: {
       commit: sinon.spy(),
       getters: {
-        'search/activeView': 'grid',
         ...mocks.$store?.getters
       },
       state: {
@@ -467,19 +466,6 @@ describe('components/search/SearchInterface', () => {
         });
       });
     });
-
-    describe('view', () => {
-      describe('setter', () => {
-        it('commits to the search store', () => {
-          const wrapper = factory();
-          const view = 'list';
-
-          wrapper.vm.view = view;
-
-          expect(wrapper.vm.$store.commit.calledWith('search/setView', view)).toBe(true);
-        });
-      });
-    });
   });
 
   describe('destroyed', () => {
@@ -729,53 +715,6 @@ describe('components/search/SearchInterface', () => {
 
             expect(wrapper.vm.$fetch.called).toBe(false);
           });
-        });
-      });
-    });
-
-    describe('setViewFromRouteQuery', () => {
-      describe('with view in route query', () => {
-        const route = { query: { view: 'mosaic', query: 'sport' } };
-
-        it('updates the stored view', () => {
-          const wrapper = factory({ mocks: { $route: route } });
-          wrapper.setData({ view: 'list' });
-
-          wrapper.vm.setViewFromRouteQuery();
-
-          expect(wrapper.vm.$store.commit.calledWith('search/setView', 'mosaic')).toBe(true);
-        });
-
-        it('sets searchResultsView cookie', () => {
-          const wrapper = factory({ mocks: { $route: route, $cookies: { set: sinon.spy() } } });
-          wrapper.setData({ view: 'list' });
-
-          wrapper.vm.setViewFromRouteQuery();
-
-          expect(wrapper.vm.$cookies.set.calledWith('searchResultsView', 'mosaic')).toBe(true);
-        });
-      });
-
-      describe('without view in route query', () => {
-        const route = { query: { query: 'sport' } };
-
-        it('does not update the stored view', () => {
-          const wrapper = factory({ mocks: { $route: route } });
-          wrapper.setData({ view: 'list' });
-          sinon.resetHistory();
-
-          wrapper.vm.setViewFromRouteQuery();
-
-          expect(wrapper.vm.$store.commit.called).toBe(false);
-        });
-
-        it('does not set searchResultsView cookie', () => {
-          const wrapper = factory({ mocks: { $route: route, $cookies: { set: sinon.spy() } } });
-          wrapper.setData({ view: 'list' });
-
-          wrapper.vm.setViewFromRouteQuery();
-
-          expect(wrapper.vm.$cookies.set.called).toBe(false);
         });
       });
     });
